fix: check team and org is not null when there is an error importing repository

+++ src/modules/reports/reports.service.ts
@@ -308,11 +308,14 @@ export class ReportsService extends AutowiredService {
         // Delete report in SFTP
         this.deleteReportFromFtp(report.id)
 
-        const team: Team = await this.teamsService.getTeamById(report.team_id)
-        const organization: Organization = await this.organizationsService.getOrganizationById(team.organization_id)
-
         // Delete all indexed contents in fulltextsearch
-        this.fullTextSearchService.deleteIndexedResults(organization.sluglified_name, team.sluglified_name, report.sluglified_name, 'report')
+        const team: Team = await this.teamsService.getTeamById(report.team_id)
+        if (team) {
+            const organization: Organization = await this.organizationsService.getOrganizationById(team.organization_id)
+            if (organization) {
+                this.fullTextSearchService.deleteIndexedResults(organization.sluglified_name, team.sluglified_name, report.sluglified_name, 'report')
+            }
+        }
 
         // Delete files
         await this.filesMongoProvider.deleteMany({ report_id: reportId })
@@ -1659,9 +1662,9 @@ export class ReportsService extends AutowiredService {
             })
             const zip = new AdmZip(response.data)
             const tmpFolder: string = await this.kysoSettingsService.getValue(KysoSettingsEnum.TMP_FOLDER_PATH)
-            Logger.log("Extracting Github files to " + tmpFolder)
+            Logger.log('Extracting Github files to ' + tmpFolder)
             zip.extractAllTo(tmpFolder, true)
-            Logger.log("Extraction finished")
+            Logger.log('Extraction finished')
             Logger.log(`Moving between ${tmpFolder}/${zip.getEntries()[0].entryName} and ${extractedDir}`)
             moveSync(`${tmpFolder}/${zip.getEntries()[0].entryName}`, extractedDir, { overwrite: true })
             Logger.log(`Moving finished`)
